feat: update model

+++ packages/llm-client/src/web/init/image.ts
@@ -32,8 +32,8 @@ export const initWebVisionModel = async (
         {
           device: "webgpu",
           dtype: {
-            embed_tokens: "fp32",
-            vision_encoder: "fp32",
+            embed_tokens: "fp16",
+            vision_encoder: "fp16",
             encoder_model: "q4",
             decoder_model_merged: "q4",
           },
